dataStoreで画面遷移を追跡して権限要求を制御
