+++ app/src/main/java/com/example/getpermissions/MainActivity.kt
@@ -1,6 +1,5 @@
 package com.example.getpermissions
 
-import android.content.Context
 import android.os.Bundle
 import androidx.activity.ComponentActivity
 import androidx.activity.compose.setContent
@@ -20,28 +19,11 @@ import androidx.compose.ui.platform.LocalContext
 import androidx.compose.ui.res.stringResource
 import androidx.compose.ui.text.style.TextAlign
 import androidx.compose.ui.unit.dp
-import androidx.datastore.core.DataStore
-import androidx.datastore.preferences.core.Preferences
-import androidx.datastore.preferences.core.booleanPreferencesKey
-import androidx.datastore.preferences.core.edit
-import androidx.datastore.preferences.core.emptyPreferences
-import androidx.datastore.preferences.preferencesDataStore
 import androidx.navigation.NavController
 import androidx.navigation.compose.NavHost
 import androidx.navigation.compose.composable
 import androidx.navigation.compose.rememberNavController
 import com.example.getpermissions.ui.theme.GetPermissionsTheme
-import kotlinx.coroutines.flow.Flow
-import kotlinx.coroutines.flow.catch
-import kotlinx.coroutines.flow.first
-import kotlinx.coroutines.flow.map
-import java.io.IOException
-
-
-// At the top level of your kotlin file:
-val Context.dataStore: DataStore<Preferences> by preferencesDataStore(name = "settings")
-// DataStoreのキーを定義
-val isFirstTimeLaunch = booleanPreferencesKey("is_first_time_launch")
 
 class MainActivity : ComponentActivity() {
     override fun onCreate(savedInstanceState: Bundle?) {
@@ -53,8 +35,9 @@ class MainActivity : ComponentActivity() {
                 val context = LocalContext.current
                 // 画面読み込み時に一度だけ実行
                 LaunchedEffect(key1 = Unit) {
-                    if(!getFirstTimeLaunchValue()){
-                        setFirstTimeLaunchValue(true)
+                    showToast(context, getFirstTimeLaunchValue(context).toString())
+                    if(getFirstTimeLaunchValue(context) == 0){
+                        setFirstTimeLaunchValue(context,1)
                     }
                     if (checkLocationPermission(context)) {
                         navController.navigate("fourth_screen") {
@@ -77,29 +60,6 @@ class MainActivity : ComponentActivity() {
                 }
             }
         }
-    }
-
-    suspend fun setFirstTimeLaunchValue(value: Boolean) {
-        applicationContext.dataStore.edit { settings ->
-            settings[isFirstTimeLaunch] = value
-        }
-    }
-
-    suspend fun getFirstTimeLaunchValue(): Boolean {
-        val isFirstTimeLaunchFlow: Flow<Boolean> = applicationContext.dataStore.data
-            .catch { exception ->
-                if (exception is IOException) {
-                    emit(emptyPreferences())
-                } else {
-                    throw exception
-                }
-            }
-            .map { preferences ->
-                // No type safety.
-                preferences[isFirstTimeLaunch] ?: false
-            }
-
-        return isFirstTimeLaunchFlow.first()
     }
 }
